[NEST-FE-21]use atom to update data
