+++ app/api/profile/get/route.ts
@@ -1,4 +1,4 @@
-import axiosInterceptor, { CustomAxiosRequestConfig } from "../../axioInterceptor";
+import axiosInterceptor, { CustomAxiosRequestConfig } from "../../axiosInterceptor";
 
 export const getProfile = async () => {
     const response = await axiosInterceptor.get("/member/me", {
