+++ app/api/profile/update/route.ts
@@ -1,4 +1,4 @@
-import axiosInterceptor, { CustomAxiosRequestConfig } from "../../axioInterceptor";
+import axiosInterceptor, { CustomAxiosRequestConfig } from "../../axiosInterceptor";
 
 export const updateProfile = async (username: string, region: string, avatar: string) => {
     const response = await axiosInterceptor.put("/member/me", {
