+++ atoms/user/atom.ts
@@ -1,0 +1,8 @@
+import { atom } from "jotai";
+
+export const userAtom = atom({
+    username: "",
+    email: "",
+    region: "",
+    avatar: "/images/default_profile_image.png",
+});
